Added needsPasswordReset property to users and build 14 which will add that property to any existing users in the database. Created UserResetPage for users whose passwords need to be reset. When creating a user there will no longer be filler data, but placeholders to let the user know what to put in each input. Updated forms to have more optional props such as unique class names for labels, inputs, etc. Broke the user entry forms into separate components.

+++ src/components/content/user-controls/UserModificationPanel.tsx
@@ -29,18 +29,20 @@ export default function UserModificationPanel(props: {
         concludeAction()
         appActions.beginLoading('Deleting user...')
 
-        deleteUser(
-          user.userID,
-          () => {
-            appActions.finishLoading()
-            appActions.notify(`Successfully deleted ${user.userID}.`)
-            handleSuccessfulDeletion()
-          },
-          () => {
-            appActions.finishLoading()
-            appActions.notify(`Failed to delete ${user.userID}.`)
-          },
-        )
+        if (user.userID) {
+          deleteUser(
+            user.userID,
+            () => {
+              appActions.finishLoading()
+              appActions.notify(`Successfully deleted ${user.userID}.`)
+              handleSuccessfulDeletion()
+            },
+            () => {
+              appActions.finishLoading()
+              appActions.notify(`Failed to delete ${user.userID}.`)
+            },
+          )
+        }
       },
       {
         pendingMessageUponConfirm: 'Deleting...',
@@ -61,7 +63,11 @@ export default function UserModificationPanel(props: {
 
   let containerClassName: string = 'UserModificationPanel hidden'
 
-  if (currentUser && restrictedAccessRoles.includes(currentUser.role)) {
+  if (
+    currentUser &&
+    currentUser.role &&
+    restrictedAccessRoles.includes(currentUser.role)
+  ) {
     containerClassName = 'UserModificationPanel'
   }
